[all] created new folder messaging to contain new files FileMessage.h/.cpp
