+++ SofaKernel/framework/sofa/helper/CMakeLists.txt
@@ -125,7 +125,7 @@ set(HEADER_FILES
     logging/ClangStyleMessageFormatter.h
     logging/DefaultStyleMessageFormatter.h
     logging/ExceptionMessageHandler.h
-    logging/FileMessage.h
+    messaging/FileMessage.h
 )
 
 set(SOURCE_FILES
@@ -203,7 +203,7 @@ set(SOURCE_FILES
     logging/LoggingMessageHandler.cpp
     logging/RoutingMessageHandler.cpp
     logging/ExceptionMessageHandler.cpp
-    logging/FileMessage.cpp
+    messaging/FileMessage.cpp
 )
 
 if(CMAKE_SYSTEM_NAME STREQUAL Linux)
