+++ SofaKernel/framework/sofa/helper/io/MeshTrian.cpp
@@ -20,7 +20,7 @@
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
 #include <sofa/helper/io/MeshTrian.h>
-#include <sofa/helper/logging/FileMessage.h>
+#include <sofa/helper/messaging/FileMessage.h>
 #include <sofa/helper/system/FileRepository.h>
 #include <sofa/helper/system/Locale.h>
 #include <cstdlib>
@@ -55,7 +55,7 @@ void MeshTrian::init (std::string filename)
         return;
     }
     FILE *f = fopen(filename.c_str(), "r");
-    msg_error_when(!f)<<sofa::helper::logging::FileMessage::UnableToOpenFile(filename.c_str());
+    msg_error_when(!f)<<sofa::helper::messaging::FileMessage::UnableToOpenFile(filename.c_str());
     if (f)
     {
         readTrian (f);
